+++ __tests__/app.js
@@ -9,17 +9,17 @@ const TEST_DESCRIPTION = 'test description';
 const TEST_EMAIL = 'test@example.com';
 const TEST_LICENSE = 'MIT';
 const TEST_NAME = 'Test Name';
-const TEST_PLUGIN_NAME = 'test-plugin-name';
+const TEST_APP_NAME = 'test-app-name';
 
 const TEST_PROMPTS = {
   description: TEST_DESCRIPTION,
   email: TEST_EMAIL,
   licenses: [TEST_LICENSE],
   name: TEST_NAME,
-  pluginName: TEST_PLUGIN_NAME
+  appName: TEST_APP_NAME
 };
 
-describe('generator-studio-plugin:supportedLicense', () => {
+describe('generator-studio-app:supportedLicense', () => {
   const supportedLicense = require(TEST_GENERATOR).supportedLicense;
 
   it('handles testing for supported licenses', () => {
@@ -28,8 +28,8 @@ describe('generator-studio-plugin:supportedLicense', () => {
   });
 });
 
-describe('generator-studio-plugin:app', () => {
-  const TEST_PLUGIN_DISPLAY_NAME = 'Test Plugin Name';
+describe('generator-studio-app:app', () => {
+  const TEST_APP_DISPLAY_NAME = 'Test App Name';
 
   beforeAll(() => {
     return helpers.run(TEST_GENERATOR).withPrompts(TEST_PROMPTS);
@@ -50,21 +50,21 @@ describe('generator-studio-plugin:app', () => {
       'README.md',
       'webpack.config.js'
     ]);
-    assert.fileContent('README.md', `# ${TEST_PLUGIN_NAME}`);
+    assert.fileContent('README.md', `# ${TEST_APP_NAME}`);
     assert.fileContent('LICENSE', `${new Date().getFullYear()}`);
     assert.fileContent('LICENSE', TEST_NAME);
     assert.jsonFileContent('package.json', {
       author: { name: TEST_NAME, email: TEST_EMAIL },
       description: TEST_DESCRIPTION,
       license: TEST_LICENSE,
-      name: TEST_PLUGIN_NAME
+      name: TEST_APP_NAME
     });
     assert.jsonFileContent('manifest.json', {
       author: { name: TEST_NAME, email: TEST_EMAIL },
       description: TEST_DESCRIPTION,
-      displayName: TEST_PLUGIN_DISPLAY_NAME,
+      displayName: TEST_APP_DISPLAY_NAME,
       licenses: [TEST_LICENSE],
-      name: TEST_PLUGIN_NAME
+      name: TEST_APP_NAME
     });
   });
 });